Architectural refacto.
