+++ packages/grabette-postprocess/README.md
@@ -54,27 +54,26 @@ docker build -t pollenrobotics/oak-vslam docker/oak_vslam/
 
 ```bash
 # 1. Convert the compact recording into the per-file oak/ layout the C++ expects
-uv run python scripts/rgbd_slam/convert_episode_to_oak.py -i ~/data/dataset/episode
+uv run python scripts/pipeline/convert_episode_to_oak.py -i ~/data/dataset/episode
 
 # 2. Run offline RTAB-Map VSLAM → camera_trajectory.csv
-uv run python scripts/rgbd_slam/run_oak_slam.py -i ~/data/dataset/episode
+uv run python scripts/pipeline/run_oak_slam.py -i ~/data/dataset/episode
 
 # 3. Validate trajectories
-uv run python scripts/check_trajectory.py ~/data/dataset
+uv run python scripts/checks/check_trajectory.py ~/data/dataset
 
 # 4. Generate LeRobot dataset
-uv run python scripts/generate_dataset.py \
+uv run python scripts/pipeline/generate_dataset.py \
   -i ~/data/dataset \
   --repo_id user/dataset-name \
   --task "task description" \
   --root ~/lerobot_datasets
 
 # 5. Visualize a trajectory
-uv run python scripts/visualize_trajectory.py ~/data/dataset/episode
-uv run python scripts/rgbd_slam/visualize_rgbd_trajectory.py ~/data/dataset/episode
+uv run python scripts/visualize/visualize_rgbd_trajectory.py ~/data/dataset/episode
 
 # 6. Push to HuggingFace Hub
-uv run python scripts/push_to_hub.py \
+uv run python scripts/pipeline/push_to_hub.py \
   --repo_id user/dataset-name \
   --root ~/lerobot_datasets
 ```
@@ -98,13 +97,13 @@ frame-to-frame deltas into absolute poses, gravity-aligns the trajectory
 
 ```bash
 # Dataset health: IMU sample counts, video metadata, flag obvious problems
-uv run python scripts/check_dataset.py ~/data/dataset
+uv run python scripts/checks/check_dataset.py ~/data/dataset
 
 # Camera-IMU synchronization (optical flow vs gyro). <20ms good, >50ms bad
-uv run python scripts/check_sync.py ~/data/dataset/episode --plot sync.png
+uv run python scripts/checks/check_sync.py ~/data/dataset/episode --plot sync.png
 
 # Trajectory quality: drift, relocalization jumps, zigzagging, unrealistic motion
-uv run python scripts/check_trajectory.py ~/data/dataset -v
+uv run python scripts/checks/check_trajectory.py ~/data/dataset -v
 ```
 
 ### 3. Generate LeRobot dataset
@@ -112,7 +111,7 @@ uv run python scripts/check_trajectory.py ~/data/dataset -v
 Converts trajectories + raw data into a LeRobot v3 dataset.
 
 ```bash
-uv run python scripts/generate_dataset.py \
+uv run python scripts/pipeline/generate_dataset.py \
   --input_dir ~/data/dataset \
   --repo_id myuser/grabette-demo \
   --task "cup manipulation" \
@@ -134,7 +133,7 @@ axis-angle rotation; the gripper component is the two joint angles.
 ```bash
 huggingface-cli login      # one-time
 
-uv run python scripts/push_to_hub.py \
+uv run python scripts/pipeline/push_to_hub.py \
   --repo_id pollenrobotics/grabette-demo \
   --root ~/lerobot_datasets
   # add --private for a private repo
@@ -142,12 +141,11 @@ uv run python scripts/push_to_hub.py \
 
 ### 5. Visualize trajectory
 
-Interactive 3D visualization with [Rerun](https://rerun.io/): trajectory,
-camera frustum, video overlay, and IMU time series.
+Interactive 3D visualization with [Rerun](https://rerun.io/): the SLAM
+trajectory, the OAK-left camera feed, and IMU time series.
 
 ```bash
-uv run python scripts/visualize_trajectory.py ~/data/dataset/episode
-uv run python scripts/rgbd_slam/visualize_rgbd_trajectory.py ~/data/dataset/episode
+uv run python scripts/visualize/visualize_rgbd_trajectory.py ~/data/dataset/episode
 ```
 
 ## Project structure
@@ -161,17 +159,18 @@ grabette-postprocess/
 │   └── dataset.py       # LeRobot v3 dataset builder
 ├── docker/
 │   └── oak_vslam/       # RTAB-Map offline_vslam C++ + Dockerfile
-└── scripts/
-    ├── rgbd_slam/
+└── scripts/                           # local CLIs (mirror the HF Space pipeline)
+    ├── pipeline/                       # the pipeline stages, in order
     │   ├── convert_episode_to_oak.py   # recording → oak/ layout
-    │   ├── run_oak_slam.py             # offline OAK-D VSLAM
-    │   └── visualize_rgbd_trajectory.py
-    ├── check_dataset.py                # dataset health check
-    ├── check_sync.py                   # camera-IMU synchronization check
-    ├── check_trajectory.py             # trajectory quality validation
-    ├── generate_dataset.py             # trajectories → LeRobot v3
-    ├── push_to_hub.py                  # upload dataset to Hugging Face Hub
-    └── visualize_trajectory.py         # Rerun 3D visualization
+    │   ├── run_oak_slam.py             # offline OAK-D VSLAM → camera_trajectory.csv
+    │   ├── generate_dataset.py         # trajectories → LeRobot v3
+    │   └── push_to_hub.py              # upload dataset to Hugging Face Hub
+    ├── checks/                         # validation / QA
+    │   ├── check_dataset.py            # dataset health check
+    │   ├── check_sync.py               # camera-IMU synchronization check
+    │   └── check_trajectory.py         # trajectory quality validation
+    └── visualize/
+        └── visualize_rgbd_trajectory.py  # Rerun 3D visualization
 ```
 
 ## Hardware
