+++ packages/grabette-postprocess/scripts/pipeline/push_to_hub.py
@@ -7,7 +7,7 @@ Requires: huggingface-cli login (or HF_TOKEN env var).
 import click
 from pathlib import Path
 
-from lerobot.datasets.lerobot_dataset import LeRobotDataset
+from grabette_postprocess.dataset import push_dataset
 
 
 @click.command()
@@ -20,19 +20,8 @@ from lerobot.datasets.lerobot_dataset import LeRobotDataset
 @click.option("--tags", multiple=True, default=["lerobot", "grabette"],
               help="Dataset tags (repeatable)")
 def main(repo_id, root, private, tags):
-    root = Path(root).expanduser().absolute()
-
-    print(f"Loading dataset {repo_id} from {root}...")
-    ds = LeRobotDataset(repo_id, root=root)
-    print(f"  Episodes: {ds.num_episodes}")
-    print(f"  Frames:   {ds.num_frames}")
-
-    print(f"\nPushing to https://huggingface.co/datasets/{repo_id} ...")
-    ds.push_to_hub(
-        tags=list(tags),
-        private=private,
-    )
-    print(f"\nDone: https://huggingface.co/datasets/{repo_id}")
+    push_dataset(repo_id, Path(root).expanduser().absolute(),
+                 private=private, tags=tuple(tags))
 
 
 if __name__ == "__main__":
